batch size integer overflow fixed batch completion calculation batch completion callback

+++ engine/ui/task.h
@@ -91,7 +91,7 @@ namespace newsflash
         std::string desc;
 
         // download/transfer size in bytes (if known)
-        std::uint32_t size;
+        std::uint64_t size;
 
         // the time elapsed running the task (in seconds)
         std::uint32_t runtime;
